requirements.txt
torch
torchvision
Pillow
tqdm
pystuck
facenet_pytorch
gradio